# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of contact book app"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://nodejs-hw-mongodb-contacts-app-gdzi.onrender.com"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login user",
        "operationId": "loginUser",
        "description": "Login a user",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/login.requestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/login.auth.response"
          },
          "400": {
            "$ref": "#/components/responses/bodyValidationError.auth.response"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized.auth.response"
          },
          "404": {
            "$ref": "#/components/responses/notFoundError.auth.response"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "schemas": {
...
      "unauthorized.error": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "required": [
              "message"
            ],
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "login.requestBody": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "description": "User's email. Should be a valid email",
            "type": "string"
          },
          "password": {
            "description": "User's password.",
            "type": "string"
          }
        },
        "example": {
          "email": "nickfury@mail.com",
          "password": "qwerty-1234"
        }
      },
      "login.success": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "required": [
              "accessToken"
            ],
            "properties": {
              "accessToken": {
                "type": "string"
              }
            }
          }
        }
      },
      "badRequestError.error": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "required": [
              "message"
            ],
            "properties": {
              "message": {
                "type": "string"
              },
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "path": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "type": {
                      "type": "string"
                    },
                    "context": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "notFoundError.error": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "required": [
              "message"
            ],
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
    "examples": {
...
      "login.auth.example": {
        "value": {
          "status": 200,
          "message": "Successfully logged in an user!",
          "data": {
            "accessToken": "racd0uTrgB6BkzKWckPmMqtXMXSroXcEz/IswTQp"
          }
        }
      },
      "bodyValidationError.auth.example": {
        "value": {
          "status": 400,
          "message": "BadRequestError",
          "data": {
            "message": "Body validation error",
            "errors": [
              {
                "message": "\\\"email\\\" must be a valid email",
                "path": [
                  "email"
                ],
                "type": "string.email",
                "context": {
                  "value": "qq",
                  "invalids": [
                    "qq"
                  ],
                  "label": "email",
                  "key": "email"
                }
              }
            ]
          }
        }
      },
      "unauthorized.auth.example": {
        "value": {
          "status": 401,
          "message": "UnauthorizedError",
          "data": {
            "message": "Unauthorized"
          }
        }
      },
      "notFoundError.auth.example": {
        "value": {
          "status": 404,
          "message": "NotFoundError",
          "data": {
            "message": "User not found"
          }
        }
      },
...
    },
    "responses": {
...
      "login.auth.response": {
        "description": "Logins a user",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/login.success"
            },
            "examples": {
              "login": {
                "$ref": "#/components/examples/login.auth.example"
              }
            }
          }
        }
      },
      "bodyValidationError.auth.response": {
        "description": "Bad validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/badRequestError.error"
            },
            "examples": {
              "bodyValidationError": {
                "$ref": "#/components/examples/bodyValidationError.auth.example"
              }
            }
          }
        }
      },
      "unauthorized.auth.response": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/unauthorized.error"
            },
            "examples": {
              "bodyValidationError": {
                "$ref": "#/components/examples/unauthorized.auth.example"
              }
            }
          }
        }
      },
      "notFoundError.auth.response": {
        "description": "Not found error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/notFoundError.error"
            },
            "examples": {
              "notFoundError": {
                "$ref": "#/components/examples/notFoundError.auth.example"
              }
            }
          }
        }
      },
...
    }
  }
}